refactor: removes dynamic property page
