+++ app/src/pages/p/index.tsx
@@ -1,0 +1,24 @@
+import { GetStaticPropsContext, NextPage } from "next";
+import { i18n } from "next-i18next";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+
+import { AppLayout } from "layouts/app-layout/AppLayout";
+import { PropertyDetailsContainer } from "app/property-details/PropertyDetailsContainer";
+
+const Index: NextPage = () => (
+  <AppLayout>
+    <PropertyDetailsContainer />
+  </AppLayout>
+);
+
+export async function getStaticProps({ locale }: GetStaticPropsContext) {
+  await i18n?.reloadResources();
+
+  return {
+    props: {
+      ...(await serverSideTranslations(locale!, ["common", "head"])),
+    },
+  };
+}
+
+export default Index;
