+++ app/src/providers/near/contract/escrow-factory.types.ts
@@ -1,0 +1,3 @@
+export type EscrowFactoryMethods = {
+  get_conditional_escrow_contracts_list: () => Promise<string[]>;
+};
